+++ packages/launchpad/src/lcdapi/base.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/naming-convention */
-import { CosmosSdkTx } from "../types";
+import { WrappedStdTx } from "../tx";
 
 /**
  * The mode used to send transaction
@@ -109,7 +109,7 @@ export interface TxsResponse {
   readonly code?: number;
   readonly raw_log: string;
   readonly logs?: unknown[];
-  readonly tx: CosmosSdkTx;
+  readonly tx: WrappedStdTx;
   /** The gas limit as set by the user */
   readonly gas_wanted?: string;
   /** The gas used by the execution */