+++ packages/launchpad/src/sequence.ts
@@ -2,7 +2,7 @@ import { Secp256k1, Secp256k1Signature, Sha256 } from "@cosmjs/crypto";
 
 import { makeSignDoc, serializeSignDoc } from "./encoding";
 import { decodeSignature } from "./signature";
-import { CosmosSdkTx } from "./types";
+import { WrappedStdTx } from "./tx";
 
 /**
  * Serach for sequence s with `min` <= `s` < `upperBound` to find the sequence that was used to sign the transaction
@@ -16,7 +16,7 @@ import { CosmosSdkTx } from "./types";
  * @returns the sequence if a match was found and undefined otherwise
  */
 export async function findSequenceForSignedTx(
-  tx: CosmosSdkTx,
+  tx: WrappedStdTx,
   chainId: string,
   accountNumber: number,
   upperBound: number,