+++ packages/launchpad/src/signingcosmosclient.ts
@@ -6,7 +6,8 @@ import { buildFeeTable, FeeTable, GasLimits, GasPrice } from "./gas";
 import { BroadcastMode } from "./lcdapi";
 import { Msg, MsgSend } from "./msgs";
 import { OfflineSigner } from "./signer";
-import { StdFee, StdTx } from "./types";
+import { StdTx } from "./tx";
+import { StdFee } from "./types";
 
 /**
  * These fees are used by the higher level methods of SigningCosmosClient